pyproject.toml
[project]
name = "cfgbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atproto>=0.0.56",
    "attrs>=24.3.0",
    "cairosvg>=2.7.1",
    "gitpython>=3.1.44",
    "httpx>=0.27.2",
    "keyring>=25.6.0",
    "mastodon-py>=1.8.1",
    "more-itertools>=10.5.0",
    "mypy>=1.14.1",
    "orjson>=3.10.13",
    "pillow>=11.0.0",
    "pydantic>=2.10.4",
    "pytest>=8.3.4",
    "rich>=13.9.4",
    "ruff>=0.8.4",
    "stamina>=24.3.0",
    "structlog>=24.4.0",
    "typer>=0.15.1",
    "urllib3>=2.3.0",
]

[project.scripts]
cfgbot = "cfgbot.cfgbot:app"
collect-cfgs = "cfgbot.collector:app"

[tool.uv]
package = true

[tool.pytest.ini_options]
xfail_strict = true
addopts = "-ra -q"
testpaths = [
    "tests",
]
